requirements.txt
fire
scikit-image
tqdm
pyyaml
opencv-python
loguru
rich
timm
tabulate
ninja

torch
torchvision
triton

ultralytics